[refactor] 패키지 구조 변경. 네이밍 변경
(cherry picked from commit 0b4e1848985f8a4fa893fc87eccf5dfdce609fb3)
