+++ src/main/java/gift/Application.java
@@ -2,8 +2,10 @@ package gift;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class Application {
 	public static void main(String[] args) {
 		SpringApplication.run(Application.class, args);
