+++ src/main/java/gift/dto/request/OrderRequest.java
@@ -1,0 +1,8 @@
+package gift.dto.request;
+
+public record OrderRequest(
+	Long optionId,
+	Long quantity,
+	String message
+) {
+}
