+++ src/main/java/gift/dto/request/UserDetails.java
@@ -1,0 +1,14 @@
+package gift.dto.request;
+
+import java.util.Set;
+
+import gift.user.domain.Role;
+
+public record UserDetails(
+	Long userId,
+	Set<Role> roles
+) {
+	public static UserDetails of(Long userId, Set<Role> roles) {
+		return new UserDetails(userId, roles);
+	}
+}
