+++ src/main/java/gift/dto/response/WishProductResponse.java
@@ -1,0 +1,11 @@
+package gift.dto.response;
+
+import gift.wishList.domain.WishProduct;
+
+public record WishProductResponse(
+	ProductResponse product
+) {
+	static public WishProductResponse from(WishProduct wishProduct) {
+		return new WishProductResponse(ProductResponse.from(wishProduct.getProduct()));
+	}
+}
