--- src/main/java/gift/feat/product/domain/SearchType.java
@@ -1,18 +1,0 @@
-package gift.feat.product.domain;
-
-
-
-public enum SearchType {
-	NAME("이름"),
-	CATEGORY("카테고리");
-
-	private final String description;
-
-	SearchType(String description) {
-		this.description = description;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-}
