--- src/main/java/gift/feat/product/repository/CategoryJpaRepository.java
@@ -1,13 +1,0 @@
-package gift.feat.product.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import gift.feat.product.domain.Category;
-
-@Repository
-public interface CategoryJpaRepository extends JpaRepository<Category, Long> {
-	Optional<Category> findByName(String name);
-}
