--- src/main/java/gift/feat/user/User.java
@@ -1,65 +1,0 @@
-package gift.feat.user;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import gift.feat.product.domain.Product;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "users")
-public class User {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(nullable = false)
-	private String email;
-
-	@Column(nullable = false)
-	private String password;
-
-	@Column(nullable = false)
-	private String role;
-
-	public User() {
-	}
-
-	private User(String email, String password, String role) {
-		this.email = email;
-		this.password = password;
-		this.role = role;
-	}
-
-	public static User of( String email, String password, String role) {
-		return new User(email, password, role);
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public String getRole() {
-		return role;
-	}
-}
