--- src/main/java/gift/feat/user/controller/UserController.java
@@ -1,14 +1,0 @@
-package gift.feat.user.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/user")
-public class UserController {
-	@GetMapping("/login")
-	public String login() {
-		return "user/login";
-	}
-}
