--- src/main/java/gift/feat/wishProduct/repository/WishProductJpaRepository.java
@@ -1,16 +1,0 @@
-package gift.feat.wishProduct.repository;
-
-import java.util.List;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import gift.feat.wishProduct.domain.WishProduct;
-
-@Repository
-public interface WishProductJpaRepository extends JpaRepository<WishProduct, Long> {
-	List<WishProduct> findByUserId(Long userId);
-	Page<WishProduct> findByUserId(Long userId, Pageable pageable);
-}
