+++ src/main/resources/templates/header.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>로그인 후 메인 페이지</title>
+    <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css" rel="stylesheet">
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+</head>
+<body>
+<header class="p-3 bg-white text-black">
+    <div class="container mx-auto">
+        <ul class="flex space-x-4">
+            <li><a id="home" href="/main" class="text-black hover:text-gray-600">Home</a></li>
+            <li><a id="product" href="/product" class="text-black hover:text-gray-600">Product</a></li>
+            <li><a id="myPage" href="/myPage" class="text-black hover:text-gray-600">MyPage</a></li>
+            <li><a id="admin" href="/templates/admin" class="text-black hover:text-gray-600">Admin</a></li>
+        </ul>
+    </div>
+</header>
+</body>
+</html>
