+++ src/main/resources/templates/user/main.th.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0"?>
+<thlogic>
+    <attr sel="#header" th:replace="~{header :: header}" />
+</thlogic>
